+++ h3/sql/updates/h3--4.0.3--unreleased.sql
@@ -43,3 +43,17 @@ CREATE OPERATOR <-> (
 );
 COMMENT ON OPERATOR <-> (h3index, h3index) IS
   'Returns the distance in grid cells between the two indices (at the lowest resolution of the two).';
+
+-- new child pos functions
+-- see https://github.com/uber/h3/pull/719
+CREATE OR REPLACE FUNCTION
+    h3_cell_to_child_pos(child h3index, parentRes integer) RETURNS int8
+AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE; COMMENT ON FUNCTION
+    h3_cell_to_child_pos(child h3index, parentRes integer)
+IS 'Returns the position of the child cell within an ordered list of all children of the cells parent at the specified resolution parentRes. The order of the ordered list is the same as that returned by cellToChildren. This is the complement of childPosToCell.';
+
+CREATE OR REPLACE FUNCTION
+    h3_child_pos_to_cell(childPos int8, parent h3index, childRes int) RETURNS h3index
+AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE; COMMENT ON FUNCTION
+    h3_child_pos_to_cell(childPos int8, parent h3index, childRes int)
+IS 'Returns the child cell at a given position within an ordered list of all children of parent at the specified resolution childRes. The order of the ordered list is the same as that returned by cellToChildren. This is the complement of cellToChildPos.';